+++ lib/src/live_streaming_config.dart
@@ -27,6 +27,7 @@ class ZegoUIKitPrebuiltLiveStreamingConfig {
         bottomMenuBarConfig = ZegoBottomMenuBarConfig(
           audienceButtons: plugins?.isEmpty ?? true
               ? []
+              //  host maybe change to be an audience
               : const [ZegoMenuBarButtonName.coHostControlButton],
         ),
         memberListConfig = ZegoMemberListConfig(),
@@ -126,6 +127,9 @@ class ZegoUIKitPrebuiltLiveStreamingConfig {
   /// if you return true in the liveback, prebuilt page will quit and return to your previous page, otherwise will ignore.
   Future<bool> Function(BuildContext context)? onLeaveConfirmation;
 
+  /// customize handling on kicked out
+  Future<void> Function(String)? onKickOut;
+
   /// customize handling after leave live streaming
   VoidCallback? onLeaveLiveStreaming;
 
@@ -139,7 +143,7 @@ class ZegoUIKitPrebuiltLiveStreamingConfig {
   ///
   /// |When the user is not speaking|When the user is speaking|
   /// |--|--|
-  /// |<img src="https://doc.oa.zego.im/Pics/ZegoUIKit/Flutter/_default_avatar_nowave.jpg" width="10%">|<img src="https://doc.oa.zego.im/Pics/ZegoUIKit/Flutter/_default_avatar.jpg" width="10%">|
+  /// |<img src="https://storage.zego.im/sdk-doc/Pics/ZegoUIKit/Flutter/_default_avatar_nowave.jpg" width="10%">|<img src="https://storage.zego.im/sdk-doc/Pics/ZegoUIKit/Flutter/_default_avatar.jpg" width="10%">|
   ///
   /// If you need to display the real avatar of your user, you can use the avatarBuilder to set the user avatar builder method (set user avatar widget builder), the usage is as follows:
   ///
@@ -164,6 +168,20 @@ class ZegoUIKitPrebuiltLiveStreamingConfig {
   /// ```
   ///
   ZegoAvatarBuilder? avatarBuilder;
+
+  ///  mark is large room or not
+  ///  sendInRoomCommand will sending to everyone in the room if true
+  ///  that mean [toUserIDs] of [sendInRoomCommand] function is disabled if true
+  bool markAsLargeRoom = false;
+
+  /// if return true, will directly open the camera when received onTurnOnYourCameraRequest
+  /// default is false
+  Future<bool> Function(BuildContext context)? onTurnOnYourCameraConfirmation;
+
+  /// if return true, will directly open the camera when received onTurnOnYourMicrophoneRequest
+  /// default is false
+  Future<bool> Function(BuildContext context)?
+      onTurnOnYourMicrophoneConfirmation;
 }
 
 class ZegoPrebuiltAudioVideoViewConfig {
@@ -179,7 +197,7 @@ class ZegoPrebuiltAudioVideoViewConfig {
   bool showSoundWavesInAudioMode;
 
   /// customize your foreground of audio video view, which is the top widget of stack
-  /// <br><img src="https://doc.oa.zego.im/Pics/ZegoUIKit/Flutter/_default_avatar_nowave.jpg" width="5%">
+  /// <br><img src="https://storage.zego.im/sdk-doc/Pics/ZegoUIKit/Flutter/_default_avatar_nowave.jpg" width="5%">
   /// you can return any widget, then we will put it on top of audio video view
   ZegoAudioVideoViewForegroundBuilder? foregroundBuilder;