+++ lib/src/live_streaming_defines.dart
@@ -8,6 +8,7 @@ enum ZegoMenuBarButtonName {
   coHostControlButton,
   beautyEffectButton,
   soundEffectButton,
+  enableChatButton,
 }
 
 class ZegoDialogInfo {