+++ lib/src/live_streaming_translation.dart
@@ -14,6 +14,7 @@ class ZegoTranslationText {
   String removeCoHostButton;
   String cancelMenuDialogButton;
   String inviteCoHostButton;
+  String kickOutUserMenuDialogButton;
 
   String noHostOnline;
   String memberListTitle;
@@ -40,6 +41,7 @@ class ZegoTranslationText {
     String? removeCoHostButton,
     String? cancelMenuDialogButton,
     String? inviteCoHostButton,
+    String? kickOutUserMenuDialogButton,
     String? noHostOnline,
     String? memberListTitle,
     String? sendRequestCoHostToast,
@@ -62,6 +64,8 @@ class ZegoTranslationText {
             cancelRequestCoHostButton ?? "Cancel the application",
         removeCoHostButton = removeCoHostButton ?? "Remove the co-host",
         inviteCoHostButton = inviteCoHostButton ?? "Invite %0 to co-host",
+        kickOutUserMenuDialogButton =
+            kickOutUserMenuDialogButton ?? "kick %0 out of the room",
         cancelMenuDialogButton = cancelMenuDialogButton ?? "Cancel",
         noHostOnline = noHostOnline ?? "No host is online.",
         memberListTitle = memberListTitle ?? "Attendance",